pyproject.toml
[project]
name = "dagster_flux"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "FluxFury", email = "flux@fluxfury.com" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"
dependencies = [
    "webnews-parser",
    "httptools>=0.5.0",
    "dagster>=1.9.6,<2",
    "dagster-cloud>=1.9.6,<2",
    "uvicorn>=0.24.0",
    "dagster-postgres>=0.26.8,<0.27",
    "langchain-openai>=0.3.12,<0.4",
    "aiokafka>=0.12.0,<0.13",
    "flux-orm",
    "kafka-python>=2.2.4",
    "langchain>=0.3.25",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9",
    "dagster-webserver>=1.9.6,<2",
]

[tool.uv]

[tool.uv.sources]
webnews-parser = { git = "ssh://git@github.com/FluxFury/webnews-parser.git", branch = "master" }
flux-orm = { git = "ssh://git@github.com/FluxFury/flux-orm.git", rev = "main" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dagster]
module_name = "dagster_flux.definitions"
project_name = "dagster_flux"

[tool.setuptools.packages.find]
exclude=["dagster_flux_tests"]

[tool.ruff]
target-version = "py312"
task-tags = ["TODO", "FIXME", "NOTE"]
fix = true
show-fixes = true
preview = true
respect-gitignore = false
line-length = 88    # Same as Black.
indent-width = 4    # Same as Black.
select = [
    "F",      # PyFlakes
    "I",      # isort
    "UP",     # PyUpgrade
    "YTT",    # flake8-2020
    "ASYNC",  # flake8-async
    "ASYNC1", # flake8-trio
    "B",      # bugbear
    "C4",     # flake8-comprehensions
    "T10",    # flake8-debugger
    "EXE",    # flake8-executable
    "INP",    # flake8-no-pep420
    "PIE",    # flake8-pie
    "PYI",    # flake8-pyi
    "PT",     # flake8-pytest
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "SLF",    # flake8-self
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "INT",    # flake8-gettext
    "ERA",    # flake8-eradicate
    "PGH",    # pygrep-hooks
    "FLY",    # flynt
    "NPY",    # NumPy-specific rules
    "PERF",   # Perflint
    "FURB",   # Refurb
    "LOG",    # flake8-logging
    "RUF",    # Ruff-specific rules
    "S",      # flake8-bandit
    "T20",    # flake8-print
    "ARG",    # flake8-unused-arguments

    "DTZ",    # flake8-datetimez
    "PD",     # pandas-vet

    "C90",    # McCabe
    "ANN",    # flake8-annotations
    "FBT",    # flake8-boolean-trap
    "RET",    # flake8-return
    "PTH",    # flake8-use-pathlib
    "TD",     # flake8-todo
    "FIX",    # flake8-fixme

    "N",      # pep8-naming
    "D",      # flake8-docstrings
    "BLE",    # flake8-blind-except
    "A",      # builtins
    "SLOT",   # flake8-slot
    "PL",     # pylint

    "W292"    # newline at end of file
]
ignore = [
    "D203", "D212", "D100", "D101", "ANN201", "D104", "D106", "TD003", "FIX002"
]
exclude = [     # Exclude a variety of commonly ignored directories.
    ".bzr",
    ".direnv",
    ".eggs",
    ".env",
    ".git",
    ".git-rewrite",
    ".hg",
    ".idea",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "infra",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
    "./.venv",
    "**/.venv/**"
]
